requirements.txt
drf-nested-routers
Django>=4.0
djangorestframework
djangorestframework-simplejwt
drf-nested-routers
